pyproject.toml
[tool.poetry]
name = "musicbott-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
ffmpeg = "^1.4"
"discord.py" = "^1.7.3"
keep-alive = "^0.0.1"
PyNaCl = "^1.4.0"
keep-repl-alive = "^1.4"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
